+++ src/main/java/nextstep/subway/exception/ValidExceptionController.java
@@ -1,5 +1,6 @@
 package nextstep.subway.exception;
 
+import nextstep.subway.applicaion.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
